% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{ArtificialDistributionChange}
\alias{ArtificialDistributionChange}
\title{An artificial dataset with sudden edge shift around at entry 9.000}
\format{A \code{"data.frame"} formatted for direct use within \code{"getMIDASscore()"}}
\source{
data.frame(src = rep(1,100000), dst = c(as.integer(rexp(n = 90000, rate = .65)) + 2, as.integer(abs(rnorm(10000, 8, .7))) + 1), times = unlist(lapply(c(1:1000), rep, times=100)))
}
\usage{
data(ArtificialDistributionChange)
}
\description{
An artificial dataset with sudden edge shift around at entry 9.000
}
\examples{
data(ArtificialDistributionChange)
plot(getMIDASScore(ArtificialDistributionChange))
}
\references{
None
}
\keyword{datasets}
