% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{MIDASexample}
\alias{MIDASexample}
\title{Dataset provided by the original C++ implementation of MIDAS}
\format{A \code{data.frame} formatted for direct use within \code{"getMIDASscore()"}}
\source{
\href{https://github.com/bhatiasiddharth/MIDAS}{Github}
}
\usage{
data(MIDASexample)
}
\description{
Dataset provided by the original C++ implementation of MIDAS
}
\examples{
data(MIDASexample)
plot(getMIDASScore(head(MIDASexample,10000)))
}
\references{
Siddharth Bhatia, Bryan Hooi, Minji Yoon, Kijung Shin, Christos Faloutsos. AAAI 2020. (\href{https://www.comp.nus.edu.sg/~sbhatia/assets/pdf/midas.pdf}{AAAI})
}
\keyword{datasets}
