% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMIDASScore}
\alias{getMIDASScore}
\title{Get the MIDAS score}
\usage{
getMIDASScore(
  input,
  rows = 2L,
  buckets = 769L,
  alpha = 0.6,
  norelations = FALSE,
  undirected = FALSE
)
}
\arguments{
\item{input}{A data.frame with integer columns src, dst & times}

\item{rows}{Number of rows/hash functions. Default is 2}

\item{buckets}{Number of buckets. Default is 769}

\item{alpha}{Temporal Decay Factor. Default is 0.6}

\item{norelations}{Run Midas instead of Midas-R. Default is False}

\item{undirected}{If graph is undirected. Default is False}
}
\value{
NumericVector of MIDAS-Scores
}
\description{
Get the MIDAS score
}
